+++ test/viking/view/helpers/url_helpers/link_to.js
@@ -17,21 +17,21 @@
     
     test("linkTo(content, url, options)", function() {
         equal(
-            Viking.View.Helpers.linkTo('Example', 'http://example.com', {class: 'myclass'}),
+            Viking.View.Helpers.linkTo('Example', 'http://example.com', {'class': 'myclass'}),
             '<a class="myclass" href="http://example.com">Example</a>'
         );
     });
     
     test("linkTo(contentFunc, url, options)", function() {
         equal(
-            Viking.View.Helpers.linkTo(function () { return 'Example'; }, 'http://example.com', {class: 'myclass'}),
+            Viking.View.Helpers.linkTo(function () { return 'Example'; }, 'http://example.com', {'class': 'myclass'}),
             '<a class="myclass" href="http://example.com">Example</a>'
         );
     });
     
     test("linkTo(url, options, contentFunc)", function() {
         equal(
-            Viking.View.Helpers.linkTo('http://example.com', {class: 'myclass'}, function () { return 'Example'; }),
+            Viking.View.Helpers.linkTo('http://example.com', {'class': 'myclass'}, function () { return 'Example'; }),
             '<a class="myclass" href="http://example.com">Example</a>'
         );
     });
@@ -42,7 +42,7 @@
         
         equal(
             Viking.View.Helpers.linkTo('Model', new Workshop({id: 10})),
-            '<a href="'+window.location.origin + '/workshops/10'+'">Model</a>'
+            '<a href="'+ window.location.protocol + '//' + window.location.host + '/workshops/10'+'">Model</a>'
         );
         
         delete Workshop;
@@ -55,7 +55,7 @@
         
         equal(
             Viking.View.Helpers.linkTo(new Workshop({id: 10}), function () { return 'Example'; }),
-            '<a href="'+window.location.origin + '/workshops/10'+'">Example</a>'
+            '<a href="' + window.location.protocol + '//' + window.location.host + '/workshops/10'+'">Example</a>'
         );
         
         delete Workshop;
@@ -67,8 +67,8 @@
         workshopPath = function(m) { return '/workshops/' + m.toParam(); }
         
         equal(
-            Viking.View.Helpers.linkTo(new Workshop({id: 10}), {class: 'myclass'}, function () { return 'Example'; }),
-            '<a class="myclass" href="'+window.location.origin + '/workshops/10'+'">Example</a>'
+            Viking.View.Helpers.linkTo(new Workshop({id: 10}), {'class': 'myclass'}, function () { return 'Example'; }),
+            '<a class="myclass" href="'+ window.location.protocol + '//' + window.location.host + '/workshops/10'+'">Example</a>'
         );
         
         delete Workshop;